Linked Pivot's swing anim with his sword.
Now we don't use parenting anymore because it was more complex, and the
benefits were small. (I.e. helping flexibility.) For now we are focusing on
the smaller details. If we expand we will just have one master xcf to have
all the sprites in. We just then toggle visibility depending on what
particular animation we're making.